requirements.txt
Flask==3.1.0
PyYAML==6.0.2
requests==2.32.3
slims-python-api==6.7.0
sample-sheet==0.13.0
Werkzeug==3.1.3
python-dotenv==1.0.0

